Updates tests to 1.2
